Update and rename g_bingo.js to g_bingo2.js
Hi there, my name is David Meech.  Mad props to you and this program, our class frequents it for every class.  I wanted to propose a couple of changes for Garth Bingo 2.0.

We are not sure what Canary is but it simply doesn't get brought up, I wanted to change that one, as well as the halflife 2 t-shirt, I am under the impression it is no more.  I've added two newer relevant ones.

Thanks in advance!